moved timer progress into a module #495
