+++ gulpfile.js
@@ -117,6 +117,7 @@ const refactoredSrc = [
   "./src/js/test/caps-warning.js",
   "./src/js/test/live-acc.js",
   "./src/js/test/test-leaderboards.js",
+  "./src/js/test/timer-progress.js",
 ];
 
 //legacy files
