+++ src/js/global-dependencies.js
@@ -46,3 +46,4 @@ import * as LiveWpm from "./live-wpm";
 import * as CapsWarning from "./caps-warning";
 import * as LiveAcc from "./live-acc";
 import * as TestLeaderboards from "./test-leaderboards";
+import * as TimerProgress from "./timer-progress";
